[refactor/#4] 아토믹 디자인  패턴 적용 리팩터
