+++ app/components/atoms/LoginInput/LoginInput.style.ts
@@ -1,0 +1,21 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  loginInput: {
+    height: 35,
+    paddingVertical: 7,
+    paddingHorizontal: 10,
+    alignSelf: "stretch",
+    borderRadius: 5,
+    borderWidth: 1,
+    borderColor: "#DBDBDB",
+    backgroundColor: "#fff",
+  },
+  loginInputText: {
+    color: "#000",
+    fontSize: 12,
+    lineHeight: 21,
+    fontWeight: "400",
+    letterSpacing: -0.32,
+  },
+});
