+++ app/components/atoms/LogoCircle/LogoCircle.style.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  logoCircle: {
+    width: 195.906,
+    height: 195.906,
+    flexShrink: 0,
+    aspectRatio: 195.91 / 195.91,
+  },
+});
