+++ app/components/molecules/LoginButton/LoginButton.style.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  loginButton: {
+    backgroundColor: "#fff",
+    width: 313,
+    height: 50,
+    borderRadius: 20,
+    borderColor: "rgba(217, 217, 217, 0.80)",
+    borderWidth: 1,
+    paddingHorizontal: 130,
+    paddingVertical: 14,
+  },
+  loginButtonPressed: {
+    backgroundColor: "#084C63",
+  },
+  loginButtonText: {
+    color: "#000",
+    fontSize: 15,
+    fontWeight: "700",
+    lineHeight: 21,
+    letterSpacing: -0.32,
+  },
+});
