+++ app/components/molecules/LoginForm/LoginForm.style.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  loginForm: {
+    display: "flex",
+    width: 300,
+    flexDirection: "column",
+    alignItems: "flex-start",
+    gap: 12,
+  },
+});
