+++ app/components/molecules/SignUpPrompt/SignUpPrompt.style.ts
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  signUpText: {
+    color: "#000",
+    textAlign: "center",
+    fontSize: 14,
+    fontWeight: "400",
+    lineHeight: 21,
+    letterSpacing: -0.32,
+  },
+  signUpTextBlue: {
+    color: "#08F",
+    marginLeft: 5,
+    fontWeight: "700",
+  },
+});
